🚨 Split `utils.context.py` into its own package
TG-2031
